+++ test/test_experiment.py
@@ -5,11 +5,7 @@ import os
 import pytest
 from pytest_mock import mocker  # noqa: F401 -- flake8 doesn't see it's used as fixture
 
-from curifactory.experiment import (
-    experiments_completer,
-    params_completer,
-    run_experiment,
-)
+from curifactory.experiment import run_experiment
 from curifactory.manager import ArtifactManager
 
 # TODO: need to test that specifying no params will default to experiment_name
@@ -463,33 +459,117 @@ def test_valid_args_names_works(clear_filesystem):
     assert manager.records[0].state["sum"] == 9
 
 
-def test_experiments_completer():
-    output = experiments_completer()
-    assert output == ["basic", "subexp.example"]
+def test_single_run_many_records_are_distinct(configured_test_manager):
+    """Running an experiment that returns multiple records with different data should
+    indeed have different data in their respective states.
+
+    This is to test that stage decorator cachers aren't singleton instances.
+    """
+    run_experiment(
+        "simple_cache",
+        ["simple_cache"],
+        param_set_names=["thing1", "thing2"],
+        mngr=configured_test_manager,
+    )
+
+    assert len(configured_test_manager.records) == 2
+    assert (
+        configured_test_manager.records[0].state["my_output"]
+        != configured_test_manager.records[1].state["my_output"]
+    )
 
 
-def test_params_completer():
-    output = params_completer()
-    assert output == ["empty", "nonarrayargs", "params1", "params2", "subparams.thing"]
+def test_double_run_many_records_are_distinct(
+    configured_test_manager, configured_test_manager2
+):
+    """Running an experiment (twice) that returns multiple records with different data should
+    indeed have different data in their respective states.
+
+    This is to test that stage decorator cachers aren't singleton instances.
+    """
+    run_experiment(
+        "simple_cache",
+        ["simple_cache"],
+        param_set_names=["thing1", "thing2"],
+        mngr=configured_test_manager,
+    )
+    run_experiment(
+        "simple_cache",
+        ["simple_cache"],
+        param_set_names=["thing1", "thing2"],
+        mngr=configured_test_manager2,
+        dry=True,
+    )  # NOTE: this works (prior to fix) with no_dag=True
+
+    assert len(configured_test_manager.records) == 2
+    assert (
+        configured_test_manager.records[0].state["my_output"]
+        != configured_test_manager.records[1].state["my_output"]
+    )
+
+    for mapped_artifact in configured_test_manager2.mapped_artifacts:
+        assert mapped_artifact.cached
+
+    assert len(configured_test_manager2.records) == 2
+    assert (
+        configured_test_manager2.records[0].state["my_output"]
+        != configured_test_manager2.records[1].state["my_output"]
+    )
 
 
-def test_macos_experiment_completer(mocker):  # noqa: F811
-    """The BSD verison of grep on macOS puts './' at the beginning of returned paths,
-    we should handle this appropriately"""
+def test_single_run_many_records_are_distinct_agg(configured_test_manager):
+    """Running an experiment that returns multiple records with different data should
+    indeed have different data in their respective states.
 
-    mock = mocker.patch("subprocess.run")
-    mock.return_value.stdout = b"./basic.py\n./subexp/example.py\n"
+    This is to test that stage decorator cachers aren't singleton instances.
+    """
+    run_experiment(
+        "simple_cache",
+        ["simple_cache"],
+        param_set_names=["thing3", "thing4"],
+        mngr=configured_test_manager,
+    )
 
-    output = experiments_completer()
-    assert output == ["basic", "subexp.example"]
+    assert len(configured_test_manager.records) == 2
+    assert (
+        configured_test_manager.records[0].state["my_agg_output"]
+        != configured_test_manager.records[1].state["my_agg_output"]
+    )
 
 
-def test_macos_params_completer(mocker):  # noqa: F811
-    """The BSD verison of grep on macOS puts './' at the beginning of returned paths,
-    we should handle this appropriately"""
+def test_double_run_many_records_are_distinct_agg(
+    configured_test_manager, configured_test_manager2
+):
+    """Running an experiment (twice) that returns multiple records with different data should
+    indeed have different data in their respective states.
 
-    mock = mocker.patch("subprocess.run")
-    mock.return_value.stdout = b"./empty.py\n./subparams/thing.py\n"
+    This is to test that stage decorator cachers aren't singleton instances.
+    """
+    run_experiment(
+        "simple_cache",
+        ["simple_cache"],
+        param_set_names=["thing3", "thing4"],
+        mngr=configured_test_manager,
+    )
+    run_experiment(
+        "simple_cache",
+        ["simple_cache"],
+        param_set_names=["thing3", "thing4"],
+        mngr=configured_test_manager2,
+        dry=True,
+    )  # NOTE: this works (prior to fix) with no_dag=True
 
-    output = params_completer()
-    assert output == ["empty", "empty", "subparams.thing", "subparams.thing"]
+    assert len(configured_test_manager.records) == 2
+    assert (
+        configured_test_manager.records[0].state["my_agg_output"]
+        != configured_test_manager.records[1].state["my_agg_output"]
+    )
+
+    for mapped_artifact in configured_test_manager2.mapped_artifacts:
+        assert mapped_artifact.cached
+
+    assert len(configured_test_manager2.records) == 2
+    assert (
+        configured_test_manager2.records[0].state["my_agg_output"]
+        != configured_test_manager2.records[1].state["my_agg_output"]
+    )